Initial commit
The initial motivation for this module is to offer a "toHtml" function
needed for the milrd case where morloc commands will be sent to the
server, compiled, run, and should return HTML blocks that will render on
the UI.

Every command, whatever its output, should be rendered in HTML. So I
need a function of type `a -> HTML`. Where `HTML` is a representable as
a string.

I really REALLY need typeclasses. But in lieu of that, I can implement a
python function that catches the main cases. It will work for now.